+++ backend/src/users/controllers/users/users.controller.ts
@@ -174,6 +174,18 @@ export class UsersController {
     return await this.userService.updateUserStatus(user.id, status);
   }
 
+  @Get('match/:id')
+  async match(@Param('id') id: number) {
+    return await this.userService.getMatch(id);
+  }
+  
+  @Get('rank/:id')
+  async rank(@Param('id') id:number) {
+      let matchs = await this.userService.getMatch(id);
+      const num = matchs.length();
+      return num;
+  }
+
   // @Patch(':id')
   // updateUserById(@Param('id', ParseIntPipe) id: number) {
   //   this.userService.updateUserById(id);
