+++ backend/src/users/services/user/user.interface.ts
@@ -11,6 +11,7 @@ export interface IUserService {
   deleteFriend(userId: number, unFriendUserId: number);
   getFriendList(userId: number);
   updateUserStatus(userId: number, status: Status);
+  getMatch(userId: number);
   // updateUserById(id: number);
   // createUser(user: UserDto);
   // deleteUser();
